+++ e2e/jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  testEnvironment: 'node',
+  testMatch: ['<rootDir>/specs/**/*.spec.js'],
+  setupFilesAfterEnv: ['<rootDir>/setup.js'],
+  testTimeout: 120000,
+  maxWorkers: 1,
+  verbose: true
+}