+++ e2e/setup.js
@@ -1,0 +1,1 @@
+jest.setTimeout(Number(process.env.E2E_TIMEOUT_MS || 120000))